+++ api.txt
@@ -1,48 +1,74 @@
 package:qr/qr.dart:
+  fromDataAndValidation (function: QrValidationResult Function({required String data, required QrErrorCorrectLevel errorCorrectLevel, required int typeNumber}))
   InputTooLongException (class extends Object implements Exception):
     new (constructor: InputTooLongException Function(int, int))
     inputLimit (getter: int)
     message (getter: String)
     providedInput (getter: int)
     toString (method: String Function())
-  QrBitBuffer (class extends Object):
-    new (constructor: QrBitBuffer Function())
-    length (getter: int)
-    length= (setter: int)
-    [] (method: bool Function(int))
-    []= (method: void Function(int, bool))
-    getByte (method: int Function(int))
-    put (method: void Function(int, int))
-    putBit (method: void Function(bool))
   QrCode (class extends Object):
-    new (constructor: QrCode Function(int, int))
-    fromData (constructor: QrCode Function({required String data, required int errorCorrectLevel}))
-    fromUint8List (constructor: QrCode Function({required Uint8List data, required int errorCorrectLevel}))
-    dataCache (getter: List<int>)
-    errorCorrectLevel (getter: int)
+    new (constructor: QrCode Function(int, QrErrorCorrectLevel))
+    fromData (constructor: QrCode Function({required String data, required QrErrorCorrectLevel errorCorrectLevel}))
+    fromUint8List (constructor: QrCode Function({required Uint8List data, required QrErrorCorrectLevel errorCorrectLevel}))
+    errorCorrectLevel (getter: QrErrorCorrectLevel)
     moduleCount (getter: int)
     typeNumber (getter: int)
     addAlphaNumeric (method: void Function(String))
     addByteData (method: void Function(ByteData))
     addData (method: void Function(String))
+    addECI (method: void Function(QrEciValue))
     addNumeric (method: void Function(String))
-  QrErrorCorrectLevel (class extends Object):
-    H (static getter: int)
-    L (static getter: int)
-    M (static getter: int)
-    Q (static getter: int)
-    levels (static getter: List<int>)
-    getName (static method: String Function(int))
-    new (constructor: QrErrorCorrectLevel Function())
+  QrEciValue (extension type implements int):
+    ascii (static getter: QrEciValue)
+    big5 (static getter: QrEciValue)
+    eucKr (static getter: QrEciValue)
+    gb2312 (static getter: QrEciValue)
+    gbk (static getter: QrEciValue)
+    iso8859_1 (static getter: QrEciValue)
+    iso8859_10 (static getter: QrEciValue)
+    iso8859_11 (static getter: QrEciValue)
+    iso8859_13 (static getter: QrEciValue)
+    iso8859_14 (static getter: QrEciValue)
+    iso8859_15 (static getter: QrEciValue)
+    iso8859_16 (static getter: QrEciValue)
+    iso8859_2 (static getter: QrEciValue)
+    iso8859_3 (static getter: QrEciValue)
+    iso8859_4 (static getter: QrEciValue)
+    iso8859_5 (static getter: QrEciValue)
+    iso8859_6 (static getter: QrEciValue)
+    iso8859_7 (static getter: QrEciValue)
+    iso8859_8 (static getter: QrEciValue)
+    iso8859_9 (static getter: QrEciValue)
+    shiftJis (static getter: QrEciValue)
+    utf16BE (static getter: QrEciValue)
+    utf8 (static getter: QrEciValue)
+    windows1250 (static getter: QrEciValue)
+    windows1251 (static getter: QrEciValue)
+    windows1252 (static getter: QrEciValue)
+    windows1256 (static getter: QrEciValue)
+    new (constructor: QrEciValue Function(int))
+    value (getter: int)
+  QrErrorCorrectLevel (enum):
+    high (static getter: QrErrorCorrectLevel)
+    low (static getter: QrErrorCorrectLevel)
+    medium (static getter: QrErrorCorrectLevel)
+    quartile (static getter: QrErrorCorrectLevel)
+    values (static getter: List<QrErrorCorrectLevel>)
+    recoveryRate (getter: int)
   QrImage (class extends Object):
     new (constructor: QrImage Function(QrCode))
     withMaskPattern (constructor: QrImage Function(QrCode, int))
-    errorCorrectLevel (getter: int)
+    errorCorrectLevel (getter: QrErrorCorrectLevel)
     maskPattern (getter: int)
     moduleCount (getter: int)
     qrModules (getter: List<List<bool?>>)
     typeNumber (getter: int)
     isDark (method: bool Function(int, int))
+  QrValidationResult (class extends Object, final):
+    isValid (getter: bool)
+    qrCode (getter: QrCode?)
+    validErrorCorrectLevels (getter: List<QrErrorCorrectLevel>)
+    validTypeNumbers (getter: List<int>)
 dart:core:
   Exception (referenced)
   List (referenced)
