+++ lib/qr.dart
@@ -1,10 +1,7 @@
-export 'src/bit_buffer.dart';
-export 'src/byte.dart';
-export 'src/eci.dart';
 export 'src/ecivalue.dart';
 export 'src/error_correct_level.dart';
 export 'src/input_too_long_exception.dart';
-export 'src/mode.dart';
-export 'src/qr_code.dart';
+export 'src/qr_code.dart' show QrCode;
 export 'src/qr_image.dart';
-export 'src/validation_result.dart';
+export 'src/validation_result.dart'
+    show QrValidationResult, fromDataAndValidation;
