+++ lib/src/qr_code.dart
@@ -6,13 +6,13 @@ import 'package:meta/meta.dart';
 import 'bit_buffer.dart';
 import 'byte.dart';
 import 'eci.dart';
+import 'ecivalue.dart';
 import 'error_correct_level.dart';
 import 'input_too_long_exception.dart';
 import 'math.dart' as qr_math;
 
 import 'polynomial.dart';
 import 'rs_block.dart';
-import 'validation_result.dart';
 
 class QrCode {
   final int typeNumber;
@@ -57,78 +57,6 @@ class QrCode {
     return QrCode(typeNumber, errorCorrectLevel).._addToList(datum);
   }
 
-  static QrValidationResult fromDataAndValidation({
-    required String data,
-    required int typeNumber,
-    required QrErrorCorrectLevel errorCorrectLevel,
-  }) {
-    final datumList = QrDatum.toDatums(data);
-
-    int calculateRequiredBits(int type) {
-      var bits = 0;
-      for (final datum in datumList) {
-        bits += 4 + datum.mode.getLengthBits(type) + datum.bitLength;
-      }
-      return bits;
-    }
-
-    // Required bits only changes at types 10 and 27.
-    final requiredBitsFor1 = calculateRequiredBits(1);
-    final requiredBitsFor10 = calculateRequiredBits(10);
-    final requiredBitsFor27 = calculateRequiredBits(27);
-
-    int getRequiredBits(int type) {
-      if (type < 10) return requiredBitsFor1;
-      if (type < 27) return requiredBitsFor10;
-      return requiredBitsFor27;
-    }
-
-    // 1. Validate Types
-    final validTypes = <int>[];
-    for (var type = 1; type <= 40; type++) {
-      final required = getRequiredBits(type);
-      final capacity = _calculateTotalDataBits(type, errorCorrectLevel);
-      if (required <= capacity) {
-        // Found minType, all subsequent types are also valid.
-        for (var t = type; t <= 40; t++) {
-          validTypes.add(t);
-        }
-        break;
-      }
-    }
-
-    // 2. Validate Error Levels
-    final validErrorLevels = <QrErrorCorrectLevel>[];
-    for (final level in QrErrorCorrectLevel.values) {
-      final requiredForType = getRequiredBits(typeNumber);
-      final capacity = _calculateTotalDataBits(typeNumber, level);
-      if (requiredForType <= capacity) {
-        validErrorLevels.add(level);
-      }
-    }
-
-    // 3. Generate Code if valid
-    QrCode? code;
-    if (validTypes.contains(typeNumber) &&
-        validErrorLevels.contains(errorCorrectLevel)) {
-      code = QrCode(typeNumber, errorCorrectLevel);
-      for (final datum in datumList) {
-        code._addToList(datum);
-      }
-    }
-
-    return QrValidationResult(
-      qrCode: code,
-      validTypeNumbers: validTypes,
-      validErrorCorrectLevels: validErrorLevels,
-    );
-  }
-
-  static int _calculateTotalDataBits(
-    int typeNumber,
-    QrErrorCorrectLevel errorCorrectLevel,
-  ) => QrRsBlock.getTotalDataBits(typeNumber, errorCorrectLevel);
-
   static int _calculateTypeNumberFromData(
     QrErrorCorrectLevel errorCorrectLevel,
     List<QrDatum> data,
@@ -147,7 +75,7 @@ class QrCode {
     final requiredBitsFor27 = getRequiredBits(27);
 
     for (var typeNumber = 1; typeNumber <= 40; typeNumber++) {
-      final totalDataBits = _calculateTotalDataBits(
+      final totalDataBits = QrRsBlock.getTotalDataBits(
         typeNumber,
         errorCorrectLevel,
       );
@@ -164,7 +92,7 @@ class QrCode {
       if (requiredBits <= totalDataBits) return typeNumber;
     }
 
-    final maxBits = _calculateTotalDataBits(40, errorCorrectLevel);
+    final maxBits = QrRsBlock.getTotalDataBits(40, errorCorrectLevel);
     throw InputTooLongException(requiredBitsFor27, maxBits);
   }
 
@@ -186,17 +114,27 @@ class QrCode {
   void addAlphaNumeric(String alphaNumeric) =>
       _addToList(QrAlphaNumeric.fromString(alphaNumeric));
 
-  void addECI(int eciValue) => _addToList(QrEci(eciValue));
+  /// Add Extended Channel Interpretation (ECI) mode data.
+  ///
+  /// Use this to specify a different character encoding for subsequent data.
+  /// Common values are available as constants on [QrEciValue].
+  void addECI(QrEciValue eciValue) => _addToList(QrEci(eciValue));
 
   void _addToList(QrDatum data) {
     _dataList.add(data);
     _dataCache = null;
   }
-
-  @internal
-  List<int> get dataCache =>
-      _dataCache ??= _createData(typeNumber, errorCorrectLevel, _dataList);
 }
+
+@internal
+void addToList(QrCode code, QrDatum datum) => code._addToList(datum);
+
+@internal
+List<int> getDataCache(QrCode code) => code._dataCache ??= _createData(
+  code.typeNumber,
+  code.errorCorrectLevel,
+  code._dataList,
+);
 
 const int _pad0 = 0xEC;
 const int _pad1 = 0x11;
