+++ test/qr_alphanumeric_test.dart
@@ -1,5 +1,7 @@
 import 'package:characters/characters.dart';
-import 'package:qr/qr.dart';
+import 'package:qr/src/bit_buffer.dart';
+import 'package:qr/src/byte.dart';
+import 'package:qr/src/mode.dart';
 import 'package:test/test.dart';
 
 void main() {
