+++ test/qr_byte_test.dart
@@ -1,6 +1,7 @@
 import 'dart:typed_data';
 
-import 'package:qr/qr.dart';
+import 'package:qr/src/bit_buffer.dart';
+import 'package:qr/src/byte.dart';
 import 'package:test/test.dart';
 
 void main() {
