+++ test/security_vulnerability_test.dart
@@ -1,4 +1,4 @@
-import 'package:qr/qr.dart';
+import 'package:qr/src/byte.dart';
 import 'package:test/test.dart';
 
 void main() {
